+++ Update-Lbi.Tests.ps1
@@ -1,20 +1,5 @@
-﻿# Copyright (c) 2017 Sergey Mazurkin
-#
-# Licensed under the Apache License, Version 2.0 (the "License");
-# you may not use this file except in compliance with the License.
-# You may obtain a copy of the License at
-#
-# http://www.apache.org/licenses/LICENSE-2.0
-#
-# Unless required by applicable law or agreed to in writing, software
-# distributed under the License is distributed on an "AS IS" BASIS,
-# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-# See the License for the specific language governing permissions and
-# limitations under the License.
-#
-# mazzy@mazzy.ru, 2017-09-28
+﻿# mazzy@mazzy.ru, 2017-10-01
 # https://github.com/mazzy-ax/Update-Lbi
-#
 
 #requires -version 3.0
 
@@ -23,19 +8,71 @@ $here = Split-Path -Parent $MyInvocation.MyCommand.Path
 Import-Module -Force ".\Update-Lbi.psm1"
 
 Describe "Update-Lbi" {
-    $Source = Join-Path $here 'TestData\*'
-    $Target = 'TestDrive:\'
-    Copy-Item $Source $Target -Recurse
 
-    Update-LibItemsCli 'TestDrive:\*.html'
+    Context "Example 1: Simple Use" {
+        $Source = Join-Path $here 'TestData\*'
+        $Target = 'TestDrive:\'
+        Copy-Item $Source $Target -Recurse -Force
+        Push-Location $Target
 
-    It "index.html" {
-        $Index = Get-Content 'TestDrive:\index.expected.html' -Encoding UTF8 -Raw
-        Get-Content 'TestDrive:\index.html' -Encoding UTF8 -Raw | Should be $Index
+        Update-Lbi -Recurse
+
+        Pop-Location
+
+        It "index.html" {
+            $Index = Get-Content 'TestDrive:\index.expected' -Encoding UTF8 -Raw
+            Get-Content 'TestDrive:\index.html' -Encoding UTF8 -Raw | Should be $Index
+        }
+
+        It "test.html" {
+            $Test = Get-Content 'TestDrive:\test.expected' -Encoding UTF8 -Raw
+            Get-Content 'TestDrive:\test.html' -Encoding UTF8 -Raw | Should be $Test
+        }
     }
 
-    It "test.html" {
-        $Test = Get-Content 'TestDrive:\test.expected.html' -Encoding UTF8 -Raw
-        Get-Content 'TestDrive:\test.html' -Encoding UTF8 -Raw | Should be $Test
+    Context "Example 2: Update some Lbi only" {
+        $Source = Join-Path $here 'TestData\*'
+        $Target = 'TestDrive:\'
+        Copy-Item $Source $Target -Recurse -Force
+        Push-Location $Target
+
+        Read-Lbi
+        Update-Lbi -UseCachedLbiOnly
+
+        Pop-Location
+
+        It "index.html" {
+            $Index = Get-Content 'TestDrive:\index.expected' -Encoding UTF8 -Raw
+            Get-Content 'TestDrive:\index.html' -Encoding UTF8 -Raw | Should be $Index
+        }
+
+        It "test.html" {
+            $Test = Get-Content 'TestDrive:\test.expected' -Encoding UTF8 -Raw
+            Get-Content 'TestDrive:\test.html' -Encoding UTF8 -Raw | Should be $Test
+        }
+    }
+
+    Context "Example 3: Update some files only" {
+        $Source = Join-Path $here 'TestData\*'
+        $Target = 'TestDrive:\'
+        Copy-Item $Source $Target -Recurse -Force
+        Push-Location $Target
+
+        Reset-LbiCache | ForEach-Object {
+            Update-Lbi 'TestDrive:\index.html' -SkipResetLbiCache
+            Update-Lbi 'TestDrive:\test.html' -SkipResetLbiCache
+        }
+
+        Pop-Location
+
+        It "index.html" {
+            $Index = Get-Content 'TestDrive:\index.expected' -Encoding UTF8 -Raw
+            Get-Content 'TestDrive:\index.html' -Encoding UTF8 -Raw | Should be $Index
+        }
+
+        It "test.html" {
+            $Test = Get-Content 'TestDrive:\test.expected' -Encoding UTF8 -Raw
+            Get-Content 'TestDrive:\test.html' -Encoding UTF8 -Raw | Should be $Test
+        }
     }
 }